requirements/dev.txt
ipdb==0.10.1
factory-boy==2.7.0
django-webtest==1.7.9